requirements.txt
requests
pyinstaller
python-dotenv
openai